Is Xanax legal in UK? Where Can You Buy Xanax Legally in the UK?
As we all want to know is xanax legal in uk so prescription only as Xanax is not legal in the UK. But People in the UK use the online website to buy Xanax online in Uk. In the Uk, medicine is used to treat and cure anxiety but in us, it is available over the counter. However, NHS opened an investigation in the UK that people are buying Xanax illegally in the Uk And the investigation is still open on this.